- Added first implementation of JSON Decoder - Redesign test classes - Improvements in errors management

+++ src/main/java/com/easyjson/json/tokenizer/JSONTokenizer.java
@@ -262,7 +262,7 @@ public class JSONTokenizer
             try
             {
                 long l = Long.parseLong(number );
-                return Token.getToken(TokenType.INTEGER, l);
+                return Token.getToken(TokenType.LONG, l);
             }
             catch(NumberFormatException nfe)
             {
